requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
onnxruntime>=1.15.0

# Energy monitoring
codecarbon>=2.3.0
psutil>=5.9.0

# Blockchain
web3>=6.0.0
py-solc-x>=1.1.0

# Simulation
simpy>=4.0.1
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
cryptography